+++ test/src/service/v1/accounts/accounts_v1_service_test.dart
@@ -2357,4 +2357,73 @@ void main() {
       );
     });
   });
+
+  group('.lookupFavouritedStatuses', () {
+    test('normal case', () async {
+      final accountsService = AccountsV1Service(
+        instance: 'test',
+        context: context.buildGetStub(
+          'test',
+          UserContext.oauth2Only,
+          '/api/v1/favourites',
+          'test/src/service/v1/accounts/data/lookup_favourited_statuses.json',
+          {
+            'limit': '40',
+          },
+        ),
+      );
+
+      final response = await accountsService.lookupFavouritedStatuses(
+        limit: 40,
+      );
+
+      expect(response, isA<MastodonResponse>());
+      expect(response.rateLimit, isA<RateLimit>());
+      expect(response.data, isA<List<Status>>());
+    });
+
+    test('when unauthorized', () async {
+      final accountsService = AccountsV1Service(
+        instance: 'test',
+        context: context.buildGetStub(
+          'test',
+          UserContext.oauth2Only,
+          '/api/v1/favourites',
+          'test/src/service/v1/accounts/data/lookup_favourited_statuses.json',
+          {
+            'limit': '40',
+          },
+          statusCode: 401,
+        ),
+      );
+
+      expectUnauthorizedException(
+        () async => await accountsService.lookupFavouritedStatuses(
+          limit: 40,
+        ),
+      );
+    });
+
+    test('when rate limit exceeded', () async {
+      final accountsService = AccountsV1Service(
+        instance: 'test',
+        context: context.buildGetStub(
+          'test',
+          UserContext.oauth2Only,
+          '/api/v1/favourites',
+          'test/src/service/v1/accounts/data/lookup_favourited_statuses.json',
+          {
+            'limit': '40',
+          },
+          statusCode: 429,
+        ),
+      );
+
+      expectRateLimitExceededException(
+        () async => await accountsService.lookupFavouritedStatuses(
+          limit: 40,
+        ),
+      );
+    });
+  });
 }